+++ NEWS.md
@@ -3,6 +3,7 @@
 * Update data
 * Added focus section for Vercelli and Alessandria
 * Update istruzioni (ita)
+* update team group eng
 * added alessandria and vercelli
 * translated x-y labels of the graphs in the epidemic section (#4)
 * translated Novara 2020-03-20 (#20)