+++ NEWS.md
@@ -1,3 +1,6 @@
+
+* update data
+
 # covid19ita 0.16.0
 
 * Update data