+++ NEWS.md
@@ -1,5 +1,5 @@
-
 * update data
+* Update team group
 
 # covid19ita 0.16.0